+++ app/src/main/java/com/example/qrchive/Fragments/CodesFragment.java
@@ -13,10 +13,10 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
-import com.example.qrchive.MyScannedCodeCardRecyclerViewAdapter;
+import com.example.qrchive.Classes.MyScannedCodeCardRecyclerViewAdapter;
 import com.example.qrchive.R;
 
-import com.example.qrchive.ScannedCode;
+import com.example.qrchive.Classes.ScannedCode;
 import com.google.android.gms.tasks.OnCompleteListener;
 import com.google.android.gms.tasks.Task;
 import com.google.firebase.firestore.DocumentSnapshot;
